Add initial support for encrypted database fields.

+++ src/format/KeePass2XmlWriter.cpp
@@ -21,20 +21,23 @@
 #include <QtCore/QFile>
 
 #include "core/Metadata.h"
+#include "crypto/SymmetricCipher.h"
 
 KeePass2XmlWriter::KeePass2XmlWriter()
     : m_db(0)
     , m_meta(0)
+    , m_cipher(0)
 {
     m_xml.setAutoFormatting(true);
     m_xml.setAutoFormattingIndent(-1); // 1 tab
     m_xml.setCodec("UTF-8");
 }
 
-void KeePass2XmlWriter::writeDatabase(QIODevice* device, Database* db)
+void KeePass2XmlWriter::writeDatabase(QIODevice* device, Database* db, SymmetricCipher* cipher)
 {
     m_db = db;
     m_meta = db->metadata();
+    m_cipher = cipher;
 
     m_xml.setDevice(device);
 
@@ -50,11 +53,11 @@ void KeePass2XmlWriter::writeDatabase(QIODevice* device, Database* db)
     m_xml.writeEndDocument();
 }
 
-void KeePass2XmlWriter::writeDatabase(const QString& filename, Database* db)
+void KeePass2XmlWriter::writeDatabase(const QString& filename, Database* db, SymmetricCipher* cipher)
 {
     QFile file(filename);
     file.open(QIODevice::WriteOnly|QIODevice::Truncate);
-    writeDatabase(&file, db);
+    writeDatabase(&file, db, cipher);
 }
 
 void KeePass2XmlWriter::writeMetadata()
@@ -271,15 +274,50 @@ void KeePass2XmlWriter::writeEntry(const Entry* entry)
 
     Q_FOREACH (const QString& key, entry->attributes().keys()) {
         m_xml.writeStartElement("String");
-        writeString("Key", key);
-        writeString("Value", entry->attributes().value(key));
+
+        bool protect = ( ((key == "Title") && m_meta->protectTitle()) ||
+                         ((key == "UserName") && m_meta->protectUsername()) ||
+                         ((key == "Password") && m_meta->protectPassword()) ||
+                         ((key == "URL") && m_meta->protectUrl()) ||
+                         ((key == "Notes") && m_meta->protectNotes()) ||
+                         entry->isAttributeProtected(key) ) &&
+                       m_cipher;
+
+        m_xml.writeStartElement("Key");
+        if (protect) {
+            m_xml.writeAttribute("Protected", "True");
+        }
+        m_xml.writeCharacters(key);
+        m_xml.writeEndElement();
+
+        if (protect) {
+            writeBinary("Value", m_cipher->process(entry->attributes().value(key).toUtf8()));
+        }
+        else {
+            writeString("Value", entry->attributes().value(key));
+        }
+
         m_xml.writeEndElement();
     }
 
     Q_FOREACH (const QString& key, entry->attachments().keys()) {
         m_xml.writeStartElement("Binary");
-        writeString("Key", key);
-        writeBinary("Value", entry->attachments().value(key));
+
+        bool protect = entry->isAttachmentProtected(key) && m_cipher;
+        m_xml.writeStartElement("Key");
+        if (protect) {
+            m_xml.writeAttribute("Protected", "True");
+        }
+        m_xml.writeCharacters(key);
+        m_xml.writeEndElement();
+
+        if (protect) {
+            writeBinary("Value", m_cipher->process(entry->attachments().value(key)));
+        }
+        else {
+            writeBinary("Value", entry->attachments().value(key));
+        }
+
         m_xml.writeEndElement();
     }
